:recycle: Refactor: 전체적인 폴더 구조 변경
